pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docgen"
version = "0.1.0"
description = "CLI to generate standardized documentation from a local Git repository"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
]

[project.scripts]
docgen = "docgen.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["docgen*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
